Refactored and separated into files, added tests
